fix bug

+++ utils/ops.py
@@ -9,7 +9,7 @@ def rev_conv2d(outs, scope, keep_r=1.0, train=True, data_format='NHWC'):
     hw_dim = np.prod(outs.shape.as_list()[2:])
     new_shape = [-1, outs.shape.as_list()[1]] + [hw_dim]
     outs = tf.reshape(outs, new_shape, name=scope+'/reshape1')
-    num_outs = outs.shape.as_list()[-1]/2
+    num_outs = int(outs.shape.as_list()[-1]/2)
     kernel = num_outs+1
     outs = conv1d(
         outs, num_outs, kernel, scope+'/conv1d', 1, keep_r, train,
